% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apres_surf.R
\name{apres_surf}
\alias{apres_surf}
\title{Apres Surf
This function helps you decide what to pack in your cooler for the beach.}
\usage{
apres_surf(timeday, age)
}
\arguments{
\item{timeday}{must be either "morning", "noon", or "afternoon"}

\item{age}{age of the surfer}
}
\description{
Apres Surf
This function helps you decide what to pack in your cooler for the beach.
}
\examples{
apres_surf("noon", 12)
}
