% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach_gear.R
\name{beach_gear}
\alias{beach_gear}
\title{Beach Gear
This function helps you decide what to pack for the beach!}
\usage{
beach_gear(season, people)
}
\arguments{
\item{season}{must be "winter", "spring", "fall", or "summer"}

\item{people}{how many people will be joining}
}
\description{
Beach Gear
This function helps you decide what to pack for the beach!
}
\examples{
beach_gear("winter", 5)
}
