% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gone-surfing.R
\name{gone_surfing}
\alias{gone_surfing}
\title{Gone Surfing}
\usage{
gone_surfing(waves, size, temp)
}
\arguments{
\item{waves}{Can you see waves? Should be a "yes" or "no".}

\item{size}{Size of the waves (ft)}

\item{temp}{Temperature of the water (degrees F)}
}
\description{
A function to build a surf report based on observable conditions.
}
\examples{
gone_surfing("yes", 4, 72)
}
